pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "infbench"
version = "0.1.0"
description = "Benchmark submission framework for distributed serving workloads"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]

[project.scripts]
infbench = "infbench.cli.submit:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
infbench = ["slurm/templates/*.j2"]

[tool.ruff]
line-length = 120
target-version = "py310"
